chore(image): move SfImage to atoms
BREAKING CHANGE: SfImage moved to atoms
