+++ packages/vue/src/components/atoms/SfImage/SfImage.js
