+++ packages/vue/src/components/atoms/SfImage/SfImage.stories.js
